Preserve current window whilst selecting item
In Neovim, after the vim-picker window is closed the cursor does not
necessarily return to the window it was in beforehand. For example,
given a single vertical split with the cursor in the right hand window,
after calling vim-picker the cursor moves to the left window. This
causes the selected item to be opened in the wrong window.

To fix this, we now save the current window ID with `win_getid` before
opening the vim-picker window, and jump back to it with `win_gotoid`
before opening the selected item.