Ensure terminal window is closed after selection
In the case that the vim-picker window contains the only listed buffer
(for example, because the window from which vim-picker was invoked
contains an unlisted buffer such as Netrw), `:bdelete!` will not close
the vim-picker window after the selector has exited. Therefore, use
`:close!` instead of `:bdelete!` so that the window is closed.