Refactor type checking of user-set variables
Previously, the functions which checked the type of user-set
configuration variables both checked the type of the variable, and
printed an error message if the type was incorrect. In retrospect, this
is a mixing of responsibilities. Now, the functions just check the type
of the variable passed to them and return a boolean. At the call site
the value of this boolean is used to determine whether an error message
is printed.