Delete vim-picker buffers after closing window
After changing from `:bdelete!` to `:close!` to ensure the window opened
by vim-picker is closed when all other buffers are unlisted, vim-picker
buffers are hidden but not deleted in Neovim if 'hidden' is set. This
commit changes the buffer local setting `bufhidden` to `delete` before
calling `:close!`, to cause the vim-picker buffer to be deleted.