Add parentheses to <Plug> mappings
The `<Plug>PickerBuffer` was treated ambiguously due to sharing a common
prefix with `<Plug>PickerBufferTag`, which causes Vim to delay before
executing `:PickerBuffer` whilst it wait to determine if `Tag` will be
added.

Add parentheses to the mapping names to create unique sequences and
prevent delays from ambiguity. The old mappings have been deprecated but
remain for backward compatibility.