Replace cwd with an options dict in PickFile
The options dict can have two values: `cwd`, and `line_handler`. The
line handler must be a function name which take a single value as
argument; this value will be the user selection. The intent is to let
the user do some processing on the selection before it's being used by
Vim, for example removing formatting. The handler must return a dict
with three possible keys:

- `filename`, which is mandatory and must a valid filename;
- `line`, which is optional;
- `column`, which is optional.

The callback in PickFile will take advantage of the line and column if
present and jump to it after opening the file.