Choose file listing command based on target directory
Prior to this change, the file listing command was being chosen based on
the attributes of the current working directory, not on the target
directory.

This caused problems when the current working directory was a Git
repository but the target directory wasn't, as `git ls-files` would be
used outside a working tree and fail.